New -> Apply new use cases & New Ui Features

+++ database/migrations/2025_01_29_185045_create_monitor_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
         Schema::create('monitors', function (Blueprint $table) {
             $table->id();
             $table->string('url');
-            $table->tinyInteger('interval');
+            $table->integer('interval');
             $table->tinyInteger('state');
             $table->string('time_out');
             $table->timestamp('last_check')->nullable();
